runs/kit/REQ-002/requirements.txt
# Runtime dependencies
fastapi>=0.109.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# Test dependencies
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0